refactor fragment name
