+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/feature_repository/search/RepositorySearchViewModel.kt
@@ -1,7 +1,7 @@
 /*
  * Copyright © 2021 YUMEMI Inc. All rights reserved.
  */
-package jp.co.yumemi.android.code_check
+package jp.co.yumemi.android.code_check.feature_repository.search
 
 import android.content.Context
 import android.os.Parcelable
@@ -14,14 +14,15 @@ import io.ktor.client.request.get
 import io.ktor.client.request.header
 import io.ktor.client.request.parameter
 import io.ktor.client.statement.HttpResponse
+import jp.co.yumemi.android.code_check.R
 import kotlinx.coroutines.async
 import kotlinx.parcelize.Parcelize
 import org.json.JSONObject
 
 /**
- * TwoFragment で使う
+ * RepositoryDetails で使う
  */
-class OneViewModel : ViewModel() {
+class RepositorySearchViewModel : ViewModel() {
 
     // 検索結果
     suspend fun searchResults(inputText: String, context: Context): List<Item> =
